Update README.md to incude further processes in building the ES index, metadata organization, Fine-tuning, etc.
Added sections further explaining metadata organization and using various models and techniques to finesse the data to be valuable for retrieval.  Added sections outlining what we did to create the data and then actually train the GPT-4o model we use during retrieval and summarization.  Showcased the actual architectual outline of how the pipeline works when in-use by a user, token management, how we use GPTClient as an object for multi-agent pipelines, context management, and so on.  Lastly, we end with the next steps that would bring this pipeline to even more enhanced use.