Update README.md
Minor issue with the filepath for the screenshot of the fine-tuned graph.  Now fixed.  We will next need to update some of the code snippets and explanations of context management now that we've refactored and are now using the custom `ChunkManager` class where we were once just using a large self.context string as an attribute.